Update to implement Copy trait for Ident
Until now, String was used directly for the internal representation of symbols.
But that would cause extra copying when copying it.
This commit has fixed that.

As a global variable, I defined a HashMap whose key is id and whose value is symbol(String).
Then, added a struct called Symbol, so that fields only have an id.
Finally, make Ident hold a Symbol instead of the String it previously held.